+++ HPXMLtoOpenStudio/resources/hvac_sizing.rb
@@ -2425,14 +2425,27 @@ class HVACSizing
 
     distribution_system.duct_leakage_measurements.each do |m|
       next if m.duct_leakage_total_or_to_outside != HPXML::DuctLeakageToOutside
-      next unless [HPXML::DuctTypeSupply, HPXML::DuctTypeReturn].include? m.duct_type
 
-      if m.duct_leakage_units == HPXML::UnitsPercent
-        cfms[m.duct_type] += m.duct_leakage_value * system_cfm
-      elsif m.duct_leakage_units == HPXML::UnitsCFM25
-        cfms[m.duct_type] += m.duct_leakage_value
-      elsif m.duct_leakage_units == HPXML::UnitsCFM50
-        cfms[m.duct_type] += Airflow.calc_air_leakage_at_diff_pressure(0.65, m.duct_leakage_value, 50.0, 25.0)
+      if not m.duct_type.nil?
+        duct_leakage = m.duct_leakage_value
+        if m.duct_leakage_units == HPXML::UnitsPercent
+          cfms[m.duct_type] += duct_leakage * system_cfm
+        elsif m.duct_leakage_units == HPXML::UnitsCFM25
+          cfms[m.duct_type] += duct_leakage
+        elsif m.duct_leakage_units == HPXML::UnitsCFM50
+          cfms[m.duct_type] += Airflow.calc_air_leakage_at_diff_pressure(0.65, duct_leakage, 50.0, 25.0)
+        end
+      else # Split the air distribution leakage equally between supply and return side when the air distribution system leakage split between the supply and return is not measured
+        duct_leakage = m.duct_leakage_value / 2
+        [HPXML::DuctTypeSupply, HPXML::DuctTypeReturn].each do |duct_type|
+          if m.duct_leakage_units == HPXML::UnitsPercent
+            cfms[duct_type] += duct_leakage * system_cfm
+          elsif m.duct_leakage_units == HPXML::UnitsCFM25
+            cfms[duct_type] += duct_leakage
+          elsif m.duct_leakage_units == HPXML::UnitsCFM50
+            cfms[duct_type] += Airflow.calc_air_leakage_at_diff_pressure(0.65, duct_leakage, 50.0, 25.0)
+          end
+        end
       end
     end