+++ Changelog.md
@@ -41,7 +41,8 @@ __New Features__
   - Allow additional outdoor design condition inputs: `DailyTemperatureRange` and `HumidityDifference`.
   - Miscellaneous improvements.
 - Adds more error-checking for inappropriate inputs (e.g., HVAC SHR=0 or clothes washer IMEF=0).
-- Allow alternative label energy use (W) input for ceiling fans.
+- Allows alternative label energy use (W) input for ceiling fans.
+- Allows the total duct leakage to outside input. The total duct leakage to outside will be split equally between supply and return side of the air distribution system. 
 
 __Bugfixes__
 - Fixes error if using AllowIncreasedFixedCapacities=true w/ HP detailed performance data.